citimart-backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
pymongo==4.13.2
bcrypt==4.3.0
Werkzeug==3.1.3
PyJWT==2.10.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
python-dotenv==1.1.1
